% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diceRoll}
\alias{diceRoll}
\title{Simulate rolling a dice and getting the sum}
\usage{
diceRoll(sides = 6L, numDice = 2L)
}
\arguments{
\item{sides}{An integer representing the number of sides on the dice}

\item{numDice}{The number of dice of size 'sides' to be rolled}
}
\value{
A number representing the sum of the dice rolled.
}
\description{
This function rolls dice and returns each different roll of the dice. If
asked to roll two dice, this will return a numeric vector representing the
value of each dice.
}
\examples{
diceRoll(6,1)
diceRoll(6,2)
diceRoll(100,2)
}
