% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diceSum}
\alias{diceSum}
\title{Roll a dice}
\usage{
diceSum(sides, numDice)
}
\arguments{
\item{sides}{An integer representing the number of sides on the dice}

\item{numDice}{The number of dice of size 'sides' to be rolled}
}
\value{
A number.
}
\description{
This function rolls dice and returns the sum of the dice.
}
\examples{
diceSum(6,1)
diceSum(6,2)
diceSum(100,2)
}
