% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dieElephant}
\alias{dieElephant}
\title{Play a White Elephant Game}
\arguments{
\item{n}{Number of participants in the game}

\item{sides}{Number of sides the dice have}

\item{numDice}{Number of dice to be rolled}
}
\value{
A list containing the final position of the present and a vector
representing the movement of the present.
}
\description{
Simulates a White Elephant party game. The gift starts at position "1" and
each player rolls a dice. For each number rolled, the present moves around the
circle that many positions in a counter-clockwise fashion.
}
\examples{
# 10 people playing with a six-sided die
dieElephant(10, 6, 1)

# 20 people playing with two six-sided dice
dieElephant(20, 6, 2)

}
