% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leftRightElephant}
\alias{leftRightElephant}
\title{Play a White Elephant Game with a coin}
\usage{
leftRightElephant(n, pheads)
}
\arguments{
\item{n}{Number of participants in the game}

\item{pheads}{The probability of heads, or moving the present to the left. Cannot
be a `NULL` value}
}
\value{
A list containing the final position of the present and a vector
representing the movement of the present (1 is a pass to the right, and -1
is a pass to the left).
}
\description{
Simulates a White Elephant party game. The gift starts at position "1" and
proceeds to move around the circle as each person flips a coin. If heads, the
present moves to the left. If tails, the present moves to the right.
}
\details{
To further elaborate on the rules of the game, if the player in position "1"
flips tails, the player to his right is sitting in position "2". If he flips
heads, the player to the left of position "1" is sitting in the last position.
}
\examples{
# 10 people playing with a fair coin
leftRightElephant(10, 0.5)

# 20 people playing with an unfair coin
leftRightElephant(20, 0.9)

}
