% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrDieElephant}
\alias{lrDieElephant}
\title{Play a White Elephant Game with dice and a coin}
\arguments{
\item{n}{Number of participants in the game}

\item{pheads}{The probability of heads, or passing the present to the left}

\item{sides}{Number of sides the dice have}

\item{numDice}{Number of dice to be rolled}
}
\value{
A list containing the final position of the present and a vector
representing the movement of the present, where negative numbers represent
movement to the left and postive number represent movement to the right.
}
\description{
Simulates a White Elephant party game. The gift starts at position "1" and
each player rolls dice and flips a coin. If the coin is heads, the present
gets passed to the left however many seats the dice have shown. If the coin is
tails, the present gets passed to the right however many seats the dice show.
}
\examples{
# 10 people playing with a six sided dice and a fair coin
lrDieElephant(10, 0.5, 6, 1)

# 20 people playing with two six-sided dice and an unfair coin (pheads = 0.7)
lrDieElephant(20, 0.7, 6, 2)

}
