% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monopoly}
\alias{monopoly}
\title{Play a game of Monopoly under specific circumstances}
\arguments{
\item{turns}{An integer representing the max number of turns to be played}

\item{sides}{An integer representing the size of dice to be rolled. Default is a six sided dice}

\item{numDice}{The number of dice of size 'sides' to be rolled. Default is two dice.}
}
\value{
A list of two vectors representing the number of times each space was
landed on and the number of times doubles were rolled.
}
\description{
This function allows you to look at and emulate playing a game of Monopoly
with a single player. This simulation does not keep track of money or which
properties that you buy, but instead rolls dice for a specified number of
turns and returns information specifying how many times each space was landed
on by the player during the game and how many times doubles were rolled during
the game.
}
