% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMonopoly.R
\name{plot.monopoly}
\alias{plot.monopoly}
\title{Plot results of a Monopoly simulation}
\usage{
\method{plot}{monopoly}(x, ...)
}
\arguments{
\item{x}{An object of class `monopoly` or `monoSum` to be summarized}

\item{...}{Additional arguments in `plot` function}
}
\value{
A plot showing the distribution of each winning seat and how often they won.
}
\description{
A general function to plot the results of a White Elephant simulation.
}
