% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElephant.R
\name{plot.partyGames}
\alias{plot.partyGames}
\title{Plot results of a White Elephant simulation}
\usage{
\method{plot}{partyGames}(sim, ...)
}
\arguments{
\item{sim}{An object of class `elphList` or `elphSum` to be summarized}
}
\value{
A plot showing the distribution of each winning seat and how often they won.
}
\description{
A general function to plot the results of a White Elephant simulation.
}
