% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMonopoly.R
\name{simulateMonopoly}
\alias{simulateMonopoly}
\title{Simulates multiple single-player Monopoly Games}
\usage{
simulateMonopoly(
  numGames = 10,
  maxTurns = 500,
  sides = 6,
  numDice = 2,
  cores = 1
)
}
\arguments{
\item{numGames}{Number of games to be played}

\item{maxTurns}{Max number of turns to be played}

\item{sides}{Number of sides on dice}

\item{numDice}{Number of dice rolled}

\item{cores}{Number of cores to be used.Currently only rund consistently when equal to 1.}
}
\value{
A matrix of class `monopoly`
}
\description{
A general function to simulate various multiple single-player Monopoly games.
}
\details{
This function uses Open MP to parallelize code and run a few different games
at the same time.

Currently, this only runs consistently when cores = 1.
}
\seealso{
[monopoly()]
}
