% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMonopoly.R
\name{summary.monopoly}
\alias{summary.monopoly}
\title{Summarize a simulation of monopoly games}
\usage{
\method{summary}{monopoly}(object, ...)
}
\arguments{
\item{object}{An object of class `monopoly` to be summarized}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
An object of class `monopoly` containing a table of moves over a
simulation and a table of times each chair won.
}
\description{
A general function to summarize a simulation of white elephant games.
}
