% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryElephant.R
\name{summary.partyGames}
\alias{summary.partyGames}
\title{Summarize a simulation}
\usage{
\method{summary}{partyGames}(object, ...)
}
\arguments{
\item{object}{An object of class `elphList` to be summarized}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
An object of class `elphSum` containing a table of moves over a
simulation and a table of times each chair won.
}
\description{
A general function to summarize a simulation of white elephant games.
}
