% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryElephant.R
\name{summaryElephant}
\alias{summaryElephant}
\title{Summarize a simulation}
\usage{
summaryElephant(sim)
}
\arguments{
\item{sim}{An object of class `elphList` to be summarized}
}
\value{
An object of class `elphSum` containing a table of moves over a
simulation and a table of times each chair won.
}
\description{
A general function to summarize a simulation of white elephant games.
}
