% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{twoCoinElephant}
\alias{twoCoinElephant}
\title{Play a White Elephant Game with two coins}
\arguments{
\item{n}{Number of participants in the game}

\item{pheads1}{The probability of heads, or passing the present to the left if using the first coin}

\item{pheads2}{The probability of heads, or passing the present to the left if using the second coin}
}
\value{
A list containing the final position of the present and a vector
representing the movement of the present, where negative numbers represent
movement to the left and positive number represent movement to the right.
}
\description{
Simulates a White Elephant party game. The gift starts at position "1" and
each player rolls dice and flips a randomly chosen coin from two available.
If the coin is heads, the present gets passed one to the left. If tails, the
present gets passed one to the right.
}
\examples{
# 10 people playing with a fair coin and an unfair coin
twoCoinElephant(10, 0.5, 0.7)

# 20 people playing with two unfair coins (pheads1 = 0.3, pheads2 = 0.7)
twoCoinElephant(20, 0.3, 0.7)

}
